+++ flow360/component/flow360_params/flow360_params.py
@@ -94,6 +94,7 @@ from .solvers import (
     TurbulenceModelSolverLegacy,
     TurbulenceModelSolverTypes,
 )
+from .time_stepping import SteadyTimeStepping, TimeStepping, UnsteadyTimeStepping
 from .unit_system import (
     AngularVelocityType,
     AreaType,
@@ -804,88 +805,6 @@ class MeshSlidingInterface(Flow360BaseModel):
         )
 
 
-class RampCFL(Flow360BaseModel):
-    """
-    Ramp CFL for time stepping component
-    """
-
-    type: str = pd.Field("ramp", const=True)
-    initial: Optional[PositiveFloat] = pd.Field()
-    final: Optional[PositiveFloat] = pd.Field()
-    ramp_steps: Optional[int] = pd.Field(alias="rampSteps")
-    randomizer: Optional[Dict] = pd.Field()
-
-    @classmethod
-    def default_steady(cls):
-        """
-        returns default steady CFL settings
-        """
-        return cls(initial=5, final=200, ramp_steps=40)
-
-    @classmethod
-    def default_unsteady(cls):
-        """
-        returns default unsteady CFL settings
-        """
-        return cls(initial=1, final=1e6, ramp_steps=30)
-
-
-class AdaptiveCFL(Flow360BaseModel):
-    """
-    Adaptive CFL for time stepping component
-    """
-
-    type: str = pd.Field("adaptive", const=True)
-    min: Optional[PositiveFloat] = pd.Field(default=0.1)
-    max: Optional[PositiveFloat] = pd.Field(default=10000)
-    max_relative_change: Optional[PositiveFloat] = pd.Field(alias="maxRelativeChange", default=1)
-    convergence_limiting_factor: Optional[PositiveFloat] = pd.Field(
-        alias="convergenceLimitingFactor", default=0.25
-    )
-
-
-# pylint: disable=E0213
-class BaseTimeStepping(Flow360BaseModel, metaclass=ABCMeta):
-    """
-    Base class for time stepping component
-    """
-
-    max_pseudo_steps: Optional[pd.conint(gt=0, le=100000)] = pd.Field(2000, alias="maxPseudoSteps")
-    CFL: Optional[Union[RampCFL, AdaptiveCFL]] = pd.Field()
-
-    # pylint: disable=arguments-differ
-    def to_solver(self, params: Flow360Params, **kwargs) -> BaseTimeStepping:
-        """
-        returns configuration object in flow360 units system
-        """
-        return super().to_solver(params, **kwargs)
-
-    # pylint: disable=missing-class-docstring,too-few-public-methods
-    class Config(Flow360BaseModel.Config):
-        deprecated_aliases = [DeprecatedAlias(name="physical_steps", deprecated="maxPhysicalSteps")]
-
-
-class SteadyTimeStepping(BaseTimeStepping):
-    """
-    Steady time stepping component
-    """
-
-    physical_steps: Literal[1] = pd.Field(1, alias="physicalSteps", const=True)
-    time_step_size: Literal["inf"] = pd.Field("inf", alias="timeStepSize", const=True)
-
-
-class UnsteadyTimeStepping(BaseTimeStepping):
-    """
-    Unsteady time stepping component
-    """
-
-    physical_steps: Optional[PositiveInt] = pd.Field(alias="physicalSteps")
-    time_step_size: Optional[TimeType.Positive] = pd.Field(alias="timeStepSize")
-
-
-TimeStepping = Union[SteadyTimeStepping, UnsteadyTimeStepping]
-
-
 class _GenericBoundaryWrapper(Flow360BaseModel):
     v: BoundaryType
 
@@ -919,6 +838,7 @@ class Boundaries(Flow360SortableBaseModel):
     def get_subtypes(cls) -> list:
         return list(get_args(_GenericBoundaryWrapper.__fields__["v"].type_))
 
+    # pylint: disable=no-self-argument
     @pd.root_validator(pre=True)
     def validate_boundary(cls, values):
         """Validator for boundary list section
@@ -1206,6 +1126,7 @@ class VolumeZones(Flow360SortableBaseModel):
     def get_subtypes(cls) -> list:
         return list(get_args(_GenericVolumeZonesWrapper.__fields__["v"].type_))
 
+    # pylint: disable=no-self-argument
     @pd.root_validator(pre=True)
     def validate_zone(cls, values):
         """Validator for zone list section
@@ -1626,6 +1547,7 @@ class BETDisk(Flow360BaseModel):
         alias="sectionalRadiuses", displayed="Sectional radiuses"
     )
 
+    # pylint: disable=no-self-argument
     @pd.validator("alphas")
     def check_alphas_in_order(cls, alpha):
         """
@@ -1635,6 +1557,7 @@ class BETDisk(Flow360BaseModel):
             raise ValueError("BET Disk: alphas are not in increasing order")
         return alpha
 
+    # pylint: disable=no-self-argument
     @pd.root_validator()
     def check_number_of_sections(cls, values):
         """
@@ -1833,6 +1756,7 @@ class Flow360Params(Flow360BaseModel):
         allow_but_remove = ["runControl", "testControl"]
         include_hash: bool = True
 
+    # pylint: disable=no-self-argument
     @pd.root_validator
     def check_consistency_wallFunction_and_SurfaceOutput(cls, values):
         """
@@ -1856,6 +1780,7 @@ class Flow360Params(Flow360BaseModel):
             )
         return values
 
+    # pylint: disable=no-self-argument
     @pd.root_validator
     def check_consistency_DDES_volumeOutput(cls, values):
         """
@@ -1884,6 +1809,7 @@ class Flow360Params(Flow360BaseModel):
                 )
         return values
 
+    # pylint: disable=no-self-argument
     @pd.root_validator
     def check_tri_quad_boundaries(cls, values):
         """
@@ -1891,6 +1817,7 @@ class Flow360Params(Flow360BaseModel):
         """
         return _check_tri_quad_boundaries(values)
 
+    # pylint: disable=no-self-argument
     @pd.root_validator
     def check_duplicate_boundary_name(cls, values):
         """
